requirements.txt
PyMySQL==1.1.0
PyYAML==6.0.1
tqdm==4.66.1
click==8.1.7